Improved parsing functionality
Added types.h and conversions.h header files for the parsing and
utilities, useful to parse the messages we receive in JSON format.

Extension from the variable parser to handle the custom types. Each
custom type defines the expected structure of the messages we expect to
parse and construct.

Planners can now parse directly this custom types being constructed
directly. No longer need complex iterations to parse of the incoming
JSONs, as we were doing before for the messages iterating for the JSON
arrays and parsing each key individually that we needed for the
(Waypoint,Subtask,References,etc.)

The types can be also converted into the ROS Msg we need for the action
goal. Conversions are defined in "conversions.h", which aims to be a
generic conversion methods (templated) from custom types into the
requested ROS Msg.